+++ lib/src/chart/line_chart/line_chart_data.dart
@@ -1292,7 +1292,7 @@ class LineTouchTooltipData with EquatableMixin {
     double? tooltipRoundedRadius,
     EdgeInsets? tooltipPadding,
     double? tooltipMargin,
-    FLHorizontalAlignment? tooltipAlignment,
+    FLHorizontalAlignment? tooltipHorizontalAlignment,
     double? tooltipHorizontalOffset,
     double? maxContentWidth,
     GetLineTooltipItems? getTooltipItems,
@@ -1306,7 +1306,8 @@ class LineTouchTooltipData with EquatableMixin {
         tooltipPadding = tooltipPadding ??
             const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
         tooltipMargin = tooltipMargin ?? 16,
-        tooltipAlignment = tooltipAlignment ?? FLHorizontalAlignment.center,
+        tooltipHorizontalAlignment =
+            tooltipHorizontalAlignment ?? FLHorizontalAlignment.center,
         tooltipHorizontalOffset = tooltipHorizontalOffset ?? 0,
         maxContentWidth = maxContentWidth ?? 120,
         getTooltipItems = getTooltipItems ?? defaultLineTooltipItem,
@@ -1330,7 +1331,7 @@ class LineTouchTooltipData with EquatableMixin {
   final double tooltipMargin;
 
   /// Controls showing tooltip on left side, right side or center aligned with spot, default is center
-  final FLHorizontalAlignment tooltipAlignment;
+  final FLHorizontalAlignment tooltipHorizontalAlignment;
 
   /// Applies horizontal offset for showing tooltip, default is zero.
   final double tooltipHorizontalOffset;
@@ -1363,7 +1364,7 @@ class LineTouchTooltipData with EquatableMixin {
         tooltipRoundedRadius,
         tooltipPadding,
         tooltipMargin,
-        tooltipAlignment,
+        tooltipHorizontalAlignment,
         tooltipHorizontalOffset,
         maxContentWidth,
         getTooltipItems,