+++ lib/src/chart/scatter_chart/scatter_chart_data.dart
@@ -398,7 +398,7 @@ class ScatterTouchTooltipData with EquatableMixin {
     Color? tooltipBgColor,
     double? tooltipRoundedRadius,
     EdgeInsets? tooltipPadding,
-    FLHorizontalAlignment? tooltipAlignment,
+    FLHorizontalAlignment? tooltipHorizontalAlignment,
     double? tooltipHorizontalOffset,
     double? maxContentWidth,
     GetScatterTooltipItems? getTooltipItems,
@@ -410,7 +410,8 @@ class ScatterTouchTooltipData with EquatableMixin {
         tooltipRoundedRadius = tooltipRoundedRadius ?? 4,
         tooltipPadding = tooltipPadding ??
             const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-        tooltipAlignment = tooltipAlignment ?? FLHorizontalAlignment.center,
+        tooltipHorizontalAlignment =
+            tooltipHorizontalAlignment ?? FLHorizontalAlignment.center,
         tooltipHorizontalOffset = tooltipHorizontalOffset ?? 0,
         maxContentWidth = maxContentWidth ?? 120,
         getTooltipItems = getTooltipItems ?? defaultScatterTooltipItem,
@@ -430,7 +431,7 @@ class ScatterTouchTooltipData with EquatableMixin {
   final EdgeInsets tooltipPadding;
 
   /// Controls showing tooltip on left side, right side or center aligned with spot, default is center
-  final FLHorizontalAlignment tooltipAlignment;
+  final FLHorizontalAlignment tooltipHorizontalAlignment;
 
   /// Applies horizontal offset for showing tooltip, default is zero.
   final double tooltipHorizontalOffset;
@@ -459,7 +460,7 @@ class ScatterTouchTooltipData with EquatableMixin {
         tooltipBgColor,
         tooltipRoundedRadius,
         tooltipPadding,
-        tooltipAlignment,
+        tooltipHorizontalAlignment,
         tooltipHorizontalOffset,
         maxContentWidth,
         getTooltipItems,
@@ -475,7 +476,7 @@ class ScatterTouchTooltipData with EquatableMixin {
     Color? tooltipBgColor,
     double? tooltipRoundedRadius,
     EdgeInsets? tooltipPadding,
-    FLHorizontalAlignment? tooltipAlignment,
+    FLHorizontalAlignment? tooltipHorizontalAlignment,
     double? tooltipHorizontalOffset,
     double? maxContentWidth,
     GetScatterTooltipItems? getTooltipItems,
@@ -488,7 +489,8 @@ class ScatterTouchTooltipData with EquatableMixin {
       tooltipBgColor: tooltipBgColor ?? this.tooltipBgColor,
       tooltipRoundedRadius: tooltipRoundedRadius ?? this.tooltipRoundedRadius,
       tooltipPadding: tooltipPadding ?? this.tooltipPadding,
-      tooltipAlignment: tooltipAlignment ?? this.tooltipAlignment,
+      tooltipHorizontalAlignment:
+          tooltipHorizontalAlignment ?? this.tooltipHorizontalAlignment,
       tooltipHorizontalOffset:
           tooltipHorizontalOffset ?? this.tooltipHorizontalOffset,
       maxContentWidth: maxContentWidth ?? this.maxContentWidth,